Rename LedgerBookUserControl.xaml to TopLedgerBookUserControl.xaml
Ensure controller doesnt need a reference to view
