+++ BudgetAnalyser/LedgerBook/LedgerBookGridBuilderFactory.cs
@@ -3,13 +3,15 @@
 namespace BudgetAnalyser.LedgerBook;
 
 [AutoRegisterWithIoC]
-public class LedgerBookGridBuilderFactory
+public class LedgerBookGridBuilderFactory(ILogger logger)
 {
+
     public virtual ILedgerBookGridBuilder GridBuilderV2(TopLedgerBookController ledgerBookController)
     {
         return ledgerBookController is null
             ? throw new ArgumentNullException(nameof(ledgerBookController))
             : (ILedgerBookGridBuilder)new LedgerBookGridBuilderV2(
+                logger,
                 ledgerBookController.ShowTransactionsCommand,
                 ledgerBookController.ShowBankBalancesCommand,
                 ledgerBookController.ShowRemarksCommand,
