+++ BudgetAnalyser/UI/DataTemplateDictionary.xaml
@@ -39,7 +39,7 @@
     </DataTemplate>
 
     <DataTemplate DataType="{x:Type ledgerBook:TopLedgerBookController}">
-        <ledgerBook:LedgerBookUserControl />
+        <ledgerBook:TopLedgerBookUserControl />
     </DataTemplate>
 
     <DataTemplate DataType="{x:Type transactions:LoadFileController}">
